Extract signup to its own component
